Угол поворота теперь сохраняется в int.

+++ src/logic/Map.java
@@ -71,7 +71,7 @@ public class Map {
 			writer.newLine();
 			for (int i=0; i<getCount(); i++){
 				Obj obj = getObj(i);
-				writer.write(obj.x + " " + obj.y + " " + obj.direction + " " + obj.sprite.getName());
+				writer.write(obj.x + " " + obj.y + " " + (int) obj.direction + " " + obj.sprite.getName());
 				writer.newLine();
 			}
 			writer.close();
